backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
